警告を消した

+++ src/game/DRAGON_KINGDOM/Road.cpp
@@ -87,7 +87,7 @@ void Road::Draw()
 
 bool Road::GetStartOrEndPos(D3DXVECTOR3* _checkPos, D3DXVECTOR3* _outputPos, float* _outputAngleDegree, bool* _startPos)
 {
-	float length = pow((_checkPos->x - m_StartPos.x)*(_checkPos->x - m_StartPos.x) +
+	double length = pow((_checkPos->x - m_StartPos.x)*(_checkPos->x - m_StartPos.x) +
 					   (_checkPos->z - m_StartPos.z)*(_checkPos->z - m_StartPos.z), 0.5);
 
 	if (length < 3000.f)
