+++ src/game/DRAGON_KINGDOM/RoadBuilder.cpp
@@ -55,7 +55,7 @@ Road* RoadBuilder::RoadBuild()
 
 	// エリアの数
 	int NumZ = int(length / ROAD_W_SIZE);
-	int VecLength = (NumZ * ROAD_H_SIZE);
+	int VecLength = int(NumZ * ROAD_H_SIZE);
 
 
 	// StartPosからEndPosの角度をとる
@@ -64,7 +64,7 @@ Road* RoadBuilder::RoadBuild()
 	// EndPosを原点に戻して、正規化、スケーリングして、もう一度同じ場所に戻す
 	D3DXVECTOR3 Vec = m_EndPos - m_StartPos;
 	D3DXVec3Normalize(&Vec, &Vec);
-	D3DXVec3Scale(&Vec, &Vec, VecLength);
+	D3DXVec3Scale(&Vec, &Vec, float(VecLength));
 	Vec = Vec + m_StartPos;
